--- copylot/hardware/hamamatsu_camera/demo/camera_demo.py
@@ -1,5 +1,0 @@
-from copylot.hardware.hamamatsu_camera.camera import Camera
-
-if __name__ == '__main__':
-    camera = Camera()
-    camera.run(100)
