+++ copylot/hardware/orca_camera/demo/camera_demo.py
@@ -1,0 +1,5 @@
+from copylot.hardware.orca_camera.orcacamera import OrcaCamera
+
+if __name__ == '__main__':
+    camera = OrcaCamera()
+    camera.run(100)
